Fix gpureal=double bug.

+++ include/types.h
@@ -56,7 +56,7 @@ THE SOFTWARE.
 
 typedef unsigned           bigint;                              //!< Big integer type
 typedef double             real;                                //!< Real number type on CPU
-typedef float              gpureal;                             //!< Real number type on GPU
+typedef double             gpureal;                             //!< Real number type on GPU
 typedef std::complex<real> complex;                             //!< Complex number type
 typedef vec<3,real>        vect;                                //!< 3-D vector type
